Consolidate redundant SessionKeyMarshaller instances between wildfly-clustering-web-hotrod and wildfly-clustering-web-infinispan.
