+++ clustering/web/cache/src/main/java/org/wildfly/clustering/web/cache/SessionKeyMarshaller.java
@@ -20,22 +20,21 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.web.infinispan;
+package org.wildfly.clustering.web.cache;
 
 import java.io.IOException;
 
-import org.wildfly.clustering.ee.infinispan.GroupedKey;
+import org.wildfly.clustering.ee.Key;
 import org.wildfly.clustering.marshalling.protostream.FunctionalScalarMarshaller;
-import org.wildfly.clustering.web.cache.SessionIdentifierMarshaller;
 import org.wildfly.common.function.ExceptionFunction;
 
 /**
  * Generic marshaller for cache keys containing session identifiers.
  * @author Paul Ferraro
  */
-public class SessionKeyMarshaller<K extends GroupedKey<String>> extends FunctionalScalarMarshaller<K, String> {
+public class SessionKeyMarshaller<K extends Key<String>> extends FunctionalScalarMarshaller<K, String> {
 
     public SessionKeyMarshaller(Class<K> targetClass, ExceptionFunction<String, K, IOException> resolver) {
-        super(targetClass, SessionIdentifierMarshaller.INSTANCE, GroupedKey::getId, resolver);
+        super(targetClass, SessionIdentifierMarshaller.INSTANCE, Key::getId, resolver);
     }
 }
