+++ clustering/web/hotrod/src/main/java/org/wildfly/clustering/web/hotrod/session/coarse/CoarseSessionAttributeSerializationContextInitializer.java
@@ -26,7 +26,7 @@ import org.infinispan.protostream.SerializationContext;
 import org.infinispan.protostream.SerializationContextInitializer;
 import org.kohsuke.MetaInfServices;
 import org.wildfly.clustering.marshalling.protostream.AbstractSerializationContextInitializer;
-import org.wildfly.clustering.web.hotrod.SessionKeyMarshaller;
+import org.wildfly.clustering.web.cache.SessionKeyMarshaller;
 
 /**
  * @author Paul Ferraro
