+++ clustering/web/infinispan/src/main/java/org/wildfly/clustering/web/infinispan/session/SessionSerializationContextInitializer.java
@@ -27,7 +27,7 @@ import org.infinispan.protostream.SerializationContextInitializer;
 import org.kohsuke.MetaInfServices;
 import org.wildfly.clustering.marshalling.protostream.AbstractSerializationContextInitializer;
 import org.wildfly.clustering.marshalling.protostream.EnumMarshaller;
-import org.wildfly.clustering.web.infinispan.SessionKeyMarshaller;
+import org.wildfly.clustering.web.cache.SessionKeyMarshaller;
 
 /**
  * @author Paul Ferraro
